Fix format string.
SOLVES:

WARN  | [iOS] xcodebuild:  /Users/quintinwillison/code/ably/ably-cocoa/Source/ARTStatus.m:127:76: warning: values of type 'NSInteger' should not be used as format arguments; add an explicit cast to 'long' instead [-Wformat]

+++ Source/ARTStatus.m
@@ -124,7 +124,7 @@ NSInteger getStatusFromCode(NSInteger code) {
         return nil;
     }
     
-    return [@"https://help.ably.io/error/" stringByAppendingFormat:@"%ld", self.statusCode];
+    return [@"https://help.ably.io/error/" stringByAppendingFormat:@"%lld", (long long)self.statusCode];
 }
 
 - (NSString *)requestId {
